Extract standalone part
